refactor: complete deprecated architecture cleanup
Phase 1-3 Complete: Migration to nursery-centric architecture

BREAKING CHANGES:
- Delete unused QuestionListCreator component and tests
- Delete unused useErrorHandler hook and tests
- Delete unused questionListStore and related tests
- Delete unused templates.ts constants
- Rename QuestionList component to QuestionsSection for clarity
- Update imports to use nurseryStore for AppError/LoadingState types
- Remove deprecated type exports from types/index.ts
- Clean up all affected import statements and test files

The application now uses a unified nursery-centric architecture:
- HomePage: uses nurseryStore (✓ completed)
- NurseryDetailPage: uses nurseryStore (✓ completed)
- All components: use new Nursery/VisitSession types (✓ completed)
- Backward compatibility: maintained through nurseryDataStore migration APIs

Next: Phase 4 will address remaining @deprecated types and cleanup.
