+++ src/components/index.ts
@@ -1,3 +1,2 @@
 export { Layout } from './Layout';
 export { AppRouter } from './Router';
-export { QuestionListCreator } from './QuestionListCreator';
