+++ src/types/index.ts
@@ -6,20 +6,26 @@
 // データモデル型
 export type {
   Question,
-  QuestionList,
-  QuestionListStats,
   SyncState,
   CreateQuestionInput,
-  CreateQuestionListInput,
   UpdateQuestionInput,
-  UpdateQuestionListInput,
+  // Nursery中心設計の型
+  Nursery,
+  VisitSession,
+  VisitSessionStatus,
+  CreateNurseryInput,
+  UpdateNurseryInput,
+  CreateVisitSessionInput,
+  UpdateVisitSessionInput,
+  NurseryStats,
+  VisitSessionStats,
 } from './data';
 
 // バリデーション関連型
 export type { ValidationResult } from '../utils/validation';
 
 // ストア関連型
-export type { AppError, LoadingState } from '../stores/questionListStore';
+export type { AppError, LoadingState } from '../stores/nurseryStore';
 
 // サービス関連型
 export { DataStoreError } from '../services/dataStore';
